+++ packages/adp-flp-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sap-ux/adp-flp-config-sub-generator
 
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [6e4a71a]
+    - @sap-ux/adp-tooling@1.0.15
+    - @sap-ux/flp-config-inquirer@1.0.15
+
+## 1.0.14
+
+### Patch Changes
+
+- Updated dependencies [41b3908]
+    - @sap-ux/feature-toggle@1.0.2
+    - @sap-ux/axios-extension@2.0.3
+    - @sap-ux/inquirer-common@1.0.11
+    - @sap-ux/adp-tooling@1.0.14
+    - @sap-ux/fiori-generator-shared@1.0.11
+    - @sap-ux/flp-config-inquirer@1.0.14
+    - @sap-ux/system-access@1.0.3
+
 ## 1.0.13
 
 ### Patch Changes