+++ packages/fiori-freestyle-writer/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @sap-ux/fiori-freestyle-writer
 
+## 3.0.22
+
+### Patch Changes
+
+- @sap-ux/ui5-test-writer@1.1.7
+
+## 3.0.21
+
+### Patch Changes
+
+- @sap-ux/ui5-test-writer@1.1.6
+
+## 3.0.20
+
+### Patch Changes
+
+- @sap-ux/fiori-generator-shared@1.0.11
+- @sap-ux/odata-service-writer@1.0.6
+- @sap-ux/ui5-test-writer@1.1.5
+- @sap-ux/cap-config-writer@1.0.11
+
 ## 3.0.19
 
 ### Patch Changes